refactor(S-3.01): extract handle_switch to auth/switch.rs
Move handle_switch_in_memory and handle_switch out of mod.rs.
handle_switch_in_memory re-exported under #[cfg(test)] for tests.
612 unit tests pass.
